+++ apps/hub/src/app/page.tsx
@@ -1,43 +1,21 @@
-import { Suspense } from "react";
+import { WeatherCard } from "@omnibytes/weather";
 
-import { api } from "@omnibytes/trpc/server";
-
-import { AuthShowcase } from "./_components/auth-showcase";
-import {
-  CreatePostForm,
-  PostCardSkeleton,
-  PostList,
-} from "./_components/posts";
-
-export const runtime = "edge";
+const style = {
+  background: "rgba(255, 255, 255, 0.22)",
+  borderRadius: 16,
+  boxShadow: "0 4px 30px rgba(0, 0, 0, 0.1)",
+  backdropFilter: "blur(5.9px)",
+  padding: 16,
+};
 
 export default async function HomePage() {
-  // You can await this here if you don't want to show Suspense fallback below
-  const posts = api.post.all();
-
   return (
-    <main className="container h-screen py-16">
-      <div className="flex flex-col items-center justify-center gap-4">
-        <h1 className="text-5xl font-extrabold tracking-tight sm:text-[5rem]">
-          Create <span className="text-primary">T3</span> Turbo
-        </h1>
-        <AuthShowcase />
-
-        <CreatePostForm />
-        <div className="w-full max-w-2xl overflow-y-scroll">
-          <Suspense
-            fallback={
-              <div className="flex w-full flex-col gap-4">
-                <PostCardSkeleton />
-                <PostCardSkeleton />
-                <PostCardSkeleton />
-              </div>
-            }
-          >
-            <PostList posts={posts} />
-          </Suspense>
+    <div className="h-screen w-screen bg-gradient-to-r from-fuchsia-600 to-pink-600 p-12 pl-32 text-xl text-white">
+      <div className="w-1/4">
+        <div style={style}>
+          <WeatherCard />
         </div>
       </div>
-    </main>
+    </div>
   );
 }
